% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-BoundedFloatText.R
\name{BoundedFloatText}
\alias{BoundedFloatText}
\title{BoundedFloatText widget}
\usage{
BoundedFloatText(
  continuous_update = FALSE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  max = 100,
  min = 0,
  step = NULL,
  value = 0,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{continuous_update}{Update the value as the user types. If False, update on submission, e.g., pressing Enter or navigating away.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes}

\item{max}{Max value}

\item{min}{Min value}

\item{step}{Minimum step to increment the value}

\item{value}{Float value}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
BoundedFloatText widget
}
