% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Checkbox.R
\name{Checkbox}
\alias{Checkbox}
\title{Checkbox widget}
\usage{
Checkbox(
  value = FALSE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  indent = TRUE,
  style = CheckboxStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{Bool value}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes.}

\item{indent}{Indent the control to align with other controls with a description.}

\item{style}{Must inherit from \link{jupyter.widget.CheckboxStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Checkbox widget
}
