% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-ColorPicker.R
\name{ColorPicker}
\alias{ColorPicker}
\title{ColorPicker widget}
\usage{
ColorPicker(
  value = "black",
  concise = FALSE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{The color value.}

\item{concise}{Display short version with just a color selector.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ColorPicker widget
}
