% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-ColorsInput.R
\name{ColorsInput}
\alias{ColorsInput}
\title{ColorsInput widget}
\usage{
ColorsInput(
  value = list(),
  description = "",
  description_allow_html = FALSE,
  placeholder = "​",
  allow_duplicates = TRUE,
  allowed_tags = list(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{List of string tags}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{placeholder}{(undocumented)}

\item{allow_duplicates}{(undocumented)}

\item{allowed_tags}{(undocumented)}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ColorsInput widget
}
