% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CoreWidget.R
\name{CoreWidget}
\alias{CoreWidget}
\title{Core Widget}
\usage{
CoreWidget(
  `_model_module` = "@jupyter-widgets/controls",
  `_model_module_version` = "2.0.0",
  ...,
  error_call = caller_env()
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{`_model_module`}{name of the model module}

\item{`_model_module_version`}{version of the model module}
}
\description{
Core Widget
}
