% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{FloatLogSlider}
\alias{FloatLogSlider}
\title{FloatLogSlider widget}
\usage{
FloatLogSlider(
  value = 1,
  min = 0,
  max = 4,
  step = 0.1,
  base = 10,
  behavior = "drag-tap",
  continuous_update = TRUE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  orientation = "horizontal",
  readout = TRUE,
  readout_format = ".3g",
  style = SliderStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{Float value}

\item{min}{Min value for the exponent}

\item{max}{Max value for the exponent}

\item{step}{Minimum step in the exponent to increment the value}

\item{base}{Base for the logarithm}

\item{behavior}{Slider dragging behavior.}

\item{continuous_update}{Update the value of the widget as the user is holding the slider.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes}

\item{orientation}{Vertical or horizontal.}

\item{readout}{Display the current value of the slider next to it.}

\item{readout_format}{Format for the readout}

\item{style}{Must inherit from \link{jupyter.widget.SliderStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
FloatLogSlider widget
}
