% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-FloatProgress.R
\name{FloatProgress}
\alias{FloatProgress}
\title{FloatProgress widget}
\usage{
FloatProgress(
  value = 0,
  min = 0,
  max = 100,
  bar_style = "",
  description = "",
  description_allow_html = FALSE,
  orientation = "horizontal",
  style = ProgressStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{Float value}

\item{min}{Min value}

\item{max}{Max value}

\item{bar_style}{Use a predefined styling for the progress bar.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{orientation}{Vertical or horizontal.}

\item{style}{Must inherit from \link{jupyter.widget.ProgressStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
FloatProgress widget
}
