% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-HTML.R
\name{HTML}
\alias{HTML}
\title{HTML widget}
\usage{
HTML(
  value = "",
  placeholder = "​",
  description = "",
  description_allow_html = FALSE,
  style = HTMLStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{String value}

\item{placeholder}{Placeholder text to display when nothing has been typed}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{style}{Must inherit from \link{jupyter.widget.HTMLStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
HTML widget
}
