% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Style.R
\name{HTMLStyle}
\alias{HTMLStyle}
\title{HTMLStyle widget}
\usage{
HTMLStyle(
  background = NULL,
  description_width = "",
  font_size = NULL,
  text_color = NULL,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{background}{Background specifications.}

\item{description_width}{Width of the description to the side of the control.}

\item{font_size}{Text font size.}

\item{text_color}{Text color}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
HTMLStyle widget
}
