% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntSlider.R
\name{IntSlider}
\alias{IntSlider}
\title{int slider}
\description{
int slider

int slider
}
\section{Super class}{
\code{wiipoc::jupyter.widget.Widget} -> \code{jupyter.widget.IntSlider}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.IntSlider-new}{\code{IntSlider$new()}}
\item \href{#method-jupyter.widget.IntSlider-mime_bundle}{\code{IntSlider$mime_bundle()}}
\item \href{#method-jupyter.widget.IntSlider-state}{\code{IntSlider$state()}}
\item \href{#method-jupyter.widget.IntSlider-on_update}{\code{IntSlider$on_update()}}
\item \href{#method-jupyter.widget.IntSlider-update}{\code{IntSlider$update()}}
\item \href{#method-jupyter.widget.IntSlider-clone}{\code{IntSlider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-mime_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-mime_bundle}{}}}
\subsection{Method \code{mime_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$mime_bundle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-state"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-state}{}}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$state(what)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-on_update"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-on_update}{}}}
\subsection{Method \code{on_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$on_update(handler)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-update"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$update(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntSlider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntSlider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntSlider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
