% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntSlider.R
\name{IntSliderModel}
\alias{IntSliderModel}
\title{IntSlider model}
\usage{
IntSliderModel(layout = Layout(), style = IntSliderStyle(), ...)
}
\arguments{
\item{layout}{See \code{\link[=Layout]{Layout()}}}

\item{style}{See \code{\link[=IntSliderStyle]{IntSliderStyle()}}}

\item{...}{additional model parameters}
}
\description{
IntSlider model
}
