% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButtonStyle-generated.R
\name{jupyter.widget.ButtonStyle}
\alias{jupyter.widget.ButtonStyle}
\title{ButtonStyle widget}
\description{
ButtonStyle widget

ButtonStyle widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.Style]{jupyter.widgets.base::jupyter.widget.Style}} -> \code{jupyter.widget.ButtonStyle}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{button_color}}{Color of the button}

\item{\code{font_family}}{Button text font family.}

\item{\code{font_size}}{Button text font size.}

\item{\code{font_style}}{Button text font style.}

\item{\code{font_variant}}{Button text font variant.}

\item{\code{font_weight}}{Button text font weight.}

\item{\code{text_color}}{Button text color.}

\item{\code{text_decoration}}{Button text decoration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.ButtonStyle-new}{\code{jupyter.widget.ButtonStyle$new()}}
\item \href{#method-jupyter.widget.ButtonStyle-clone}{\code{jupyter.widget.ButtonStyle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.ButtonStyle-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.ButtonStyle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.ButtonStyle$new(
  button_color = NULL,
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{button_color}}{Color of the button}

\item{\code{font_family}}{Button text font family.}

\item{\code{font_size}}{Button text font size.}

\item{\code{font_style}}{Button text font style.}

\item{\code{font_variant}}{Button text font variant.}

\item{\code{font_weight}}{Button text font weight.}

\item{\code{text_color}}{Button text color.}

\item{\code{text_decoration}}{Button text decoration.}

\item{\code{...}}{See \link[jupyter.widgets.base:Style]{jupyter.widgets.base::Style}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.ButtonStyle' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.ButtonStyle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.ButtonStyle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.ButtonStyle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
