% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{jupyter.widget.IntRangeSlider}
\alias{jupyter.widget.IntRangeSlider}
\title{IntRangeSlider widget}
\description{
IntRangeSlider widget

IntRangeSlider widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget}} -> \code{jupyter.widget.IntRangeSlider}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{behavior}}{Slider dragging behavior.}

\item{\code{continuous_update}}{Update the value of the widget as the user is sliding the slider.}

\item{\code{description}}{Description of the control.}

\item{\code{description_allow_html}}{Accept HTML in the description.}

\item{\code{disabled}}{Enable or disable user changes}

\item{\code{max}}{Max value}

\item{\code{min}}{Min value}

\item{\code{orientation}}{Vertical or horizontal.}

\item{\code{readout}}{Display the current value of the slider next to it.}

\item{\code{readout_format}}{Format for the readout}

\item{\code{step}}{Minimum step that the value can take}

\item{\code{value}}{Tuple of (lower, upper) bounds}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.IntRangeSlider-new}{\code{jupyter.widget.IntRangeSlider$new()}}
\item \href{#method-jupyter.widget.IntRangeSlider-clone}{\code{jupyter.widget.IntRangeSlider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="observe"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-observe'><code>jupyter.widgets.base::jupyter.widget.Widget$observe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.DOMWidget" data-id="mime_bundle"><a href='../../jupyter.widgets.base/html/jupyter.widget.DOMWidget.html#method-jupyter.widget.DOMWidget-mime_bundle'><code>jupyter.widgets.base::jupyter.widget.DOMWidget$mime_bundle()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntRangeSlider-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntRangeSlider-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.IntRangeSlider$new(
  behavior = "drag-tap",
  continuous_update = TRUE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  max = 100L,
  min = 0L,
  orientation = "horizontal",
  readout = TRUE,
  readout_format = "d",
  step = 1L,
  value = list(0L, 1L),
  style = SliderStyle(),
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{behavior}}{Slider dragging behavior.}

\item{\code{continuous_update}}{Update the value of the widget as the user is sliding the slider.}

\item{\code{description}}{Description of the control.}

\item{\code{description_allow_html}}{Accept HTML in the description.}

\item{\code{disabled}}{Enable or disable user changes}

\item{\code{max}}{Max value}

\item{\code{min}}{Min value}

\item{\code{orientation}}{Vertical or horizontal.}

\item{\code{readout}}{Display the current value of the slider next to it.}

\item{\code{readout_format}}{Format for the readout}

\item{\code{step}}{Minimum step that the value can take}

\item{\code{value}}{Tuple of (lower, upper) bounds}

\item{\code{style}}{Must inherit from \link{jupyter.widget.SliderStyle}.}

\item{\code{...}}{See \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::DOMWidget}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.IntRangeSlider' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.IntRangeSlider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.IntRangeSlider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.IntRangeSlider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
