% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-LabelStyle.R
\name{jupyter.widget.LabelStyle}
\alias{jupyter.widget.LabelStyle}
\title{LabelStyle widget}
\description{
LabelStyle widget

LabelStyle widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.Style]{jupyter.widgets.base::jupyter.widget.Style}} -> \code{jupyter.widget.LabelStyle}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{background}}{Background specifications.}

\item{\code{description_width}}{Width of the description to the side of the control.}

\item{\code{font_family}}{Label text font family.}

\item{\code{font_size}}{Text font size.}

\item{\code{font_style}}{Label text font style.}

\item{\code{font_variant}}{Label text font variant.}

\item{\code{font_weight}}{Label text font weight.}

\item{\code{text_color}}{Text color}

\item{\code{text_decoration}}{Label text decoration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.LabelStyle-new}{\code{jupyter.widget.LabelStyle$new()}}
\item \href{#method-jupyter.widget.LabelStyle-clone}{\code{jupyter.widget.LabelStyle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="observe"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-observe'><code>jupyter.widgets.base::jupyter.widget.Widget$observe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.LabelStyle-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.LabelStyle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.LabelStyle$new(
  background = NULL,
  description_width = "",
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{background}}{Background specifications.}

\item{\code{description_width}}{Width of the description to the side of the control.}

\item{\code{font_family}}{Label text font family.}

\item{\code{font_size}}{Text font size.}

\item{\code{font_style}}{Label text font style.}

\item{\code{font_variant}}{Label text font variant.}

\item{\code{font_weight}}{Label text font weight.}

\item{\code{text_color}}{Text color}

\item{\code{text_decoration}}{Label text decoration.}

\item{\code{...}}{See \link[jupyter.widgets.base:Style]{jupyter.widgets.base::Style}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.LabelStyle' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.LabelStyle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.LabelStyle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.LabelStyle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
