% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-SliderStyle.R
\name{jupyter.widget.SliderStyle}
\alias{jupyter.widget.SliderStyle}
\title{SliderStyle widget}
\description{
SliderStyle widget

SliderStyle widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.Style]{jupyter.widgets.base::jupyter.widget.Style}} -> \code{jupyter.widget.SliderStyle}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{description_width}}{Width of the description to the side of the control.}

\item{\code{handle_color}}{Color of the slider handle.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.SliderStyle-new}{\code{jupyter.widget.SliderStyle$new()}}
\item \href{#method-jupyter.widget.SliderStyle-clone}{\code{jupyter.widget.SliderStyle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.SliderStyle-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.SliderStyle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.SliderStyle$new(
  description_width = "",
  handle_color = NULL,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description_width}}{Width of the description to the side of the control.}

\item{\code{handle_color}}{Color of the slider handle.}

\item{\code{...}}{See \link[jupyter.widgets.base:Style]{jupyter.widgets.base::Style}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.SliderStyle' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.SliderStyle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.SliderStyle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.SliderStyle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
