% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{jupyter.widget.Stack}
\alias{jupyter.widget.Stack}
\title{Stack widget}
\description{
Stack widget

Stack widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget}} -> \code{jupyter.widget.Stack}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{children}}{List of widget children}

\item{\code{box_style}}{Use a predefined styling for the box.}

\item{\code{selected_index}}{The index of the selected page. This is either an integer selecting a particular sub-widget, or None to have no widgets selected.}

\item{\code{titles}}{Titles of the pages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.Stack-new}{\code{jupyter.widget.Stack$new()}}
\item \href{#method-jupyter.widget.Stack-clone}{\code{jupyter.widget.Stack$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="observe"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-observe'><code>jupyter.widgets.base::jupyter.widget.Widget$observe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.DOMWidget" data-id="mime_bundle"><a href='../../jupyter.widgets.base/html/jupyter.widget.DOMWidget.html#method-jupyter.widget.DOMWidget-mime_bundle'><code>jupyter.widgets.base::jupyter.widget.DOMWidget$mime_bundle()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Stack-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Stack-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Stack$new(
  children = list(),
  box_style = "",
  selected_index = NULL,
  titles = list(),
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{children}}{List of widget children}

\item{\code{box_style}}{Use a predefined styling for the box.}

\item{\code{selected_index}}{The index of the selected page. This is either an integer selecting a particular sub-widget, or None to have no widgets selected.}

\item{\code{titles}}{Titles of the pages}

\item{\code{...}}{See \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::DOMWidget}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.Stack' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Stack-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Stack-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Stack$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
