% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{jupyter.widget.Time}
\alias{jupyter.widget.Time}
\title{Time widget}
\description{
Time widget

Time widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget}} -> \code{jupyter.widget.Time}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{}

\item{\code{min}}{}

\item{\code{max}}{}

\item{\code{step}}{The time step to use for the picker, in seconds, or "any".}

\item{\code{description}}{Description of the control.}

\item{\code{description_allow_html}}{Accept HTML in the description.}

\item{\code{disabled}}{Enable or disable user changes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.Time-new}{\code{jupyter.widget.Time$new()}}
\item \href{#method-jupyter.widget.Time-clone}{\code{jupyter.widget.Time$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="observe"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-observe'><code>jupyter.widgets.base::jupyter.widget.Widget$observe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.DOMWidget" data-id="mime_bundle"><a href='../../jupyter.widgets.base/html/jupyter.widget.DOMWidget.html#method-jupyter.widget.DOMWidget-mime_bundle'><code>jupyter.widgets.base::jupyter.widget.DOMWidget$mime_bundle()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Time-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Time-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Time$new(
  value = NULL,
  min = NULL,
  max = NULL,
  step = 60L,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(undocumented)}

\item{\code{min}}{(undocumented)}

\item{\code{max}}{(undocumented)}

\item{\code{step}}{The time step to use for the picker, in seconds, or "any".}

\item{\code{description}}{Description of the control.}

\item{\code{description_allow_html}}{Accept HTML in the description.}

\item{\code{disabled}}{Enable or disable user changes.}

\item{\code{...}}{See \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::DOMWidget}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'jupyter.widget.Time' widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Time-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Time-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Time$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
