% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-LabelStyle.R
\name{LabelStyle}
\alias{LabelStyle}
\title{LabelStyle widget}
\usage{
LabelStyle(
  background = NULL,
  description_width = "",
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{background}{Background specifications.}

\item{description_width}{Width of the description to the side of the control.}

\item{font_family}{Label text font family.}

\item{font_size}{Text font size.}

\item{font_style}{Label text font style.}

\item{font_variant}{Label text font variant.}

\item{font_weight}{Label text font weight.}

\item{text_color}{Text color}

\item{text_decoration}{Label text decoration.}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
LabelStyle widget
}
