% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{NaiveDatetime}
\alias{NaiveDatetime}
\title{NaiveDatetime widget}
\usage{
NaiveDatetime(
  value = NULL,
  min = NULL,
  max = NULL,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{(undocumented)}

\item{min}{(undocumented)}

\item{max}{(undocumented)}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
NaiveDatetime widget
}
