% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Style.R
\name{ProgressStyle}
\alias{ProgressStyle}
\title{ProgressStyle widget}
\usage{
ProgressStyle(
  bar_color = NULL,
  description_width = "",
  ...,
  error_call = current_env()
)
}
\arguments{
\item{bar_color}{Color of the progress bar.}

\item{description_width}{Width of the description to the side of the control.}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ProgressStyle widget
}
