% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{SelectionSlider}
\alias{SelectionSlider}
\title{SelectionSlider widget}
\usage{
SelectionSlider(
  options = list(),
  index = 0L,
  behavior = "drag-tap",
  continuous_update = TRUE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  orientation = "horizontal",
  readout = TRUE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{options}{The labels for the options.}

\item{index}{Selected index}

\item{behavior}{Slider dragging behavior.}

\item{continuous_update}{Update the value of the widget as the user is holding the slider.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes}

\item{orientation}{Vertical or horizontal.}

\item{readout}{Display the current selected label next to the slider}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
SelectionSlider widget
}
