% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SliderStyle-generated.R, R/wiipoc.R
\name{SliderStyle}
\alias{SliderStyle}
\alias{IntSliderStyle}
\title{SliderStyle widget}
\usage{
SliderStyle(
  description_width = "",
  handle_color = NULL,
  ...,
  error_call = current_env()
)

IntSliderStyle(
  description_width = "",
  handle_color = NULL,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{description_width}{Width of the description to the side of the control.}

\item{handle_color}{Color of the slider handle.}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
SliderStyle widget
}
