% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Stack.R
\name{Stack}
\alias{Stack}
\title{Stack widget}
\usage{
Stack(
  children = list(),
  box_style = "",
  selected_index = NULL,
  titles = list(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{children}{List of widget children}

\item{box_style}{Use a predefined styling for the box.}

\item{selected_index}{The index of the selected page. This is either an integer selecting a particular sub-widget, or None to have no widgets selected.}

\item{titles}{Titles of the pages}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Stack widget
}
