% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Textarea.R
\name{Textarea}
\alias{Textarea}
\title{Textarea widget}
\usage{
Textarea(
  value = "",
  rows = NULL,
  placeholder = "​",
  continuous_update = TRUE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{String value}

\item{rows}{The number of rows to display.}

\item{placeholder}{Placeholder text to display when nothing has been typed}

\item{continuous_update}{Update the value as the user types. If False, update on submission, e.g., pressing Enter or navigating away.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Textarea widget
}
