% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{ToggleButton}
\alias{ToggleButton}
\title{ToggleButton widget}
\usage{
ToggleButton(
  value = FALSE,
  button_style = "",
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  icon = "",
  style = ToggleButtonStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{Bool value}

\item{button_style}{Use a predefined styling for the button.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes.}

\item{icon}{Font-awesome icon.}

\item{style}{Must inherit from \link{jupyter.widget.ToggleButtonStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ToggleButton widget
}
