% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Style.R
\name{ToggleButtonStyle}
\alias{ToggleButtonStyle}
\title{ToggleButtonStyle widget}
\usage{
ToggleButtonStyle(
  description_width = "",
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{description_width}{Width of the description to the side of the control.}

\item{font_family}{Toggle button text font family.}

\item{font_size}{Toggle button text font size.}

\item{font_style}{Toggle button text font style.}

\item{font_variant}{Toggle button text font variant.}

\item{font_weight}{Toggle button text font weight.}

\item{text_color}{Toggle button text color}

\item{text_decoration}{Toggle button text decoration.}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ToggleButtonStyle widget
}
