% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-VBox.R
\name{VBox}
\alias{VBox}
\title{VBox widget}
\usage{
VBox(children = list(), box_style = "", ..., error_call = current_env())
}
\arguments{
\item{children}{List of widget children}

\item{box_style}{Use a predefined styling for the box.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
VBox widget
}
