% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_widgets.R
\name{Widget}
\alias{Widget}
\title{Widget}
\usage{
Widget(..., error_call = caller_env())
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Widget
}
