% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box.R
\name{Box}
\alias{Box}
\title{Box}
\usage{
Box(
  children = list(),
  box_style = "",
  description = "Click Me",
  disabled = FALSE,
  layout = Layout(),
  style = NULL,
  tabbable = FALSE,
  tooltip = "",
  `_dom_classes` = character(),
  `_model_module` = "@jupyter-widgets/controls",
  `_model_module_version` = "2.0.0",
  `_model_name` = "BoxModel",
  `_view_module` = "@jupyter-widgets/controls",
  `_view_count` = NULL,
  `_view_module_version` = "2.0.0",
  `_view_name` = "BoxView",
  ...,
  error_call = current_env()
)
}
\arguments{
\item{children}{children widgets}

\item{box_style}{box style. empty (default) or one of 'success', 'info', 'warning', 'danger', 'primary'.}

\item{description}{text description of the button}

\item{disabled}{TRUE if the Button is disabled}

\item{layout}{a \code{\link[jupyter.widgets.base:Layout]{Layout()}}}

\item{style}{a \code{\link[jupyter.widgets.base:Style]{Style()}}}

\item{tabbable}{is the widget tabbable}

\item{tooltip}{tooltip}

\item{_dom_classes}{CSS classes applied to widget DOM element}

\item{_model_module}{The namespace of the model.}

\item{_model_module_version}{A semver requirement for namespace version containing the model.}

\item{_model_name}{model name}

\item{_view_module}{view mmodule}

\item{_view_count}{view count}

\item{_view_module_version}{view module version}

\item{_view_name}{view name}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Box
}
