% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Button.R
\name{Button}
\alias{Button}
\title{Button}
\description{
Button

Button
}
\section{Super class}{
\code{wiipoc::jupyter.widget.Widget} -> \code{jupyter.widget.Button}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.Button-new}{\code{Button$new()}}
\item \href{#method-jupyter.widget.Button-mime_bundle}{\code{Button$mime_bundle()}}
\item \href{#method-jupyter.widget.Button-state}{\code{Button$state()}}
\item \href{#method-jupyter.widget.Button-on_click}{\code{Button$on_click()}}
\item \href{#method-jupyter.widget.Button-clone}{\code{Button$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Button-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Button-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Button$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Button-mime_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Button-mime_bundle}{}}}
\subsection{Method \code{mime_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Button$mime_bundle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Button-state"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Button-state}{}}}
\subsection{Method \code{state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Button$state(what)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Button-on_click"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Button-on_click}{}}}
\subsection{Method \code{on_click()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Button$on_click(handler)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Button-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Button-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Button$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
