% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Button.R
\name{Button}
\alias{Button}
\title{Button}
\usage{
Button(layout = Layout(), style = ButtonStyle(), description = "Click Me", ...)
}
\arguments{
\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{style}{a \code{\link[=ButtonStyle]{ButtonStyle()}}}

\item{description}{text description of the button}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Button
}
