% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Button.R
\name{Button}
\alias{Button}
\title{Button}
\usage{
Button(
  layout = Layout(),
  style = ButtonStyle(),
  description = "Click Me",
  button_style = "",
  icon = "",
  ...
)
}
\arguments{
\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{style}{a \code{\link[=ButtonStyle]{ButtonStyle()}}}

\item{description}{text description of the button}

\item{button_style}{"", "primary", "success", "info", "warning" or "danger"}

\item{icon}{name of a font-awesome icon, see \code{\link[fontawesome:fa]{fontawesome::fa()}} or "" for no icon (default)}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Button
}
