% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Button.R
\name{Button}
\alias{Button}
\title{Button}
\usage{
Button(
  description = "Click Me",
  disabled = FALSE,
  button_style = "",
  icon = "",
  layout = Layout(),
  style = ButtonStyle(),
  tabbable = FALSE,
  tooltip = "",
  `_model_module` = "@jupyter-widgets/controls",
  `_model_module_version` = "2.0.0",
  `_model_name` = "ButtonModel",
  `_view_module` = "@jupyter-widgets/controls",
  `_view_count` = NULL,
  `_view_module_version` = "2.0.0",
  `_view_name` = "ButtonView",
  ...,
  error_call = current_env()
)
}
\arguments{
\item{description}{text description of the button}

\item{disabled}{TRUE if the Button is disabled}

\item{button_style}{"", "primary", "success", "info", "warning" or "danger"}

\item{icon}{name of a font-awesome icon, see \code{\link[fontawesome:fa]{fontawesome::fa()}} or "" for no icon (default)}

\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{style}{a \code{\link[=Style]{Style()}}}

\item{tabbable}{is the widget tabbable}

\item{tooltip}{tooltip}

\item{_model_module}{The namespace of the model.}

\item{_model_module_version}{A semver requirement for namespace version containing the model.}

\item{_model_name}{model name}

\item{_view_module}{view mmodule}

\item{_view_count}{view count}

\item{_view_module_version}{view module version}

\item{_view_name}{view name}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Button
}
