% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{Button}
\alias{Button}
\title{Button widget}
\usage{
Button(
  description = "",
  button_style = "",
  disabled = FALSE,
  icon = "",
  style = ButtonStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{description}{Button label.}

\item{button_style}{Use a predefined styling for the button.}

\item{disabled}{Enable or disable user changes.}

\item{icon}{Font-awesome icon names, without the 'fa-' prefix.}

\item{style}{Must inherit from \link{jupyter.widget.ButtonStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Button widget
}
