% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Button.R
\name{ButtonModel}
\alias{ButtonModel}
\title{Button Model}
\usage{
ButtonModel(layout = Layout(), style = ButtonStyle(), ...)
}
\arguments{
\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{style}{a \code{\link[=ButtonStyle]{ButtonStyle()}}}

\item{...}{additional model parameters, currently unused}
}
\description{
Button Model
}
