% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-Style.R
\name{ButtonStyle}
\alias{ButtonStyle}
\title{ButtonStyle widget}
\usage{
ButtonStyle(
  button_color = NULL,
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{button_color}{Color of the button}

\item{font_family}{Button text font family.}

\item{font_size}{Button text font size.}

\item{font_style}{Button text font style.}

\item{font_variant}{Button text font variant.}

\item{font_weight}{Button text font weight.}

\item{text_color}{Button text color.}

\item{text_decoration}{Button text decoration.}

\item{...}{forwarded to \link[jupyter.widgets.base:Style]{jupyter.widgets.base::jupyter.widget.Style} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
ButtonStyle widget
}
