% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ButtonStyle.R
\name{jupyter.widget.ButtonStyle}
\alias{jupyter.widget.ButtonStyle}
\alias{ButtonStyle}
\title{ButtonStyle widget}
\usage{
ButtonStyle(..., error_call = current_env())
}
\arguments{
\item{...}{See constructor for \code{jupyter.widgets.ButtonStyle}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a \link{jupyter.widget.ButtonStyle} object suitable for the
\code{style} argument of \code{\link[=Button]{Button()}}.
}
\description{
Configure the appearance of a \link{jupyter.widget.Button} widget
}
\section{Super classes}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{\link[jupyter.widgets.base:jupyter.widget.Style]{jupyter.widgets.base::jupyter.widget.Style}} -> \code{jupyter.widget.ButtonStyle}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{button_color}}{CSS button color}

\item{\code{font_family}}{CSS \code{font-family} property}

\item{\code{font_size}}{CSS \code{font-size} property}

\item{\code{font_style}}{CSS \code{font-style} property}

\item{\code{font_variant}}{CSS \code{font-variant} property}

\item{\code{text_color}}{CSS color for the text}

\item{\code{text_decoration}}{CSS \code{text-decoration} property}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.ButtonStyle-new}{\code{jupyter.widget.ButtonStyle$new()}}
\item \href{#method-jupyter.widget.ButtonStyle-clone}{\code{jupyter.widget.ButtonStyle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_custom"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_custom'><code>jupyter.widgets.base::jupyter.widget.Widget$on_custom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_update'><code>jupyter.widgets.base::jupyter.widget.Widget$on_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.ButtonStyle-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.ButtonStyle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.ButtonStyle$new(
  button_color = NULL,
  font_family = NULL,
  font_size = NULL,
  font_style = NULL,
  font_variant = NULL,
  font_weight = NULL,
  text_color = NULL,
  text_decoration = NULL,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{button_color}}{A valid css color, e.g. "red" or "#ff0000"}

\item{\code{font_family}}{css "font-family" value, e.g. "monospace", "Times New Roman", "Georgia, 'Times New Roman', serif", ...}

\item{\code{font_size}}{css "font-size" value, e.g. "small", "large", "16px", "2em", "80\%"}

\item{\code{font_style}}{"normal", "italic" or "oblique"}

\item{\code{font_variant}}{text formatting variations, e.g. "normal", "small-caps", ...}

\item{\code{font_weight}}{boldness of the text, either one of "normal", "bold", "lighter", and "bolder" or a number between 100 and 900}

\item{\code{text_color}}{A valid css color for the text of the Button.}

\item{\code{text_decoration}}{text styling effects, e.g. "underline, "overline", "line-through" or "blink"}

\item{\code{...}}{See \link[jupyter.widgets.base:Style]{jupyter.widgets.base::Style}}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.ButtonStyle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.ButtonStyle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.ButtonStyle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
