% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_DOMWidget.R
\name{DOMWidget}
\alias{DOMWidget}
\title{A DOM Widget}
\usage{
DOMWidget(
  layout = Layout(),
  style = NULL,
  tabbable = FALSE,
  tooltip = "",
  `_model_module` = "@jupyter-widgets/controls",
  `_model_module_version` = "2.0.0",
  ...,
  error_call = caller_env()
)
}
\arguments{
\item{layout}{a \code{\link[=Layout]{Layout()}}}

\item{style}{a \code{\link[=Style]{Style()}}}

\item{tabbable}{is the widget tabbable}

\item{tooltip}{tooltip}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
A DOM Widget
}
