% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-DOM.R
\name{IntSlider}
\alias{IntSlider}
\title{IntSlider widget}
\usage{
IntSlider(
  value = 0L,
  min = 0L,
  max = 100L,
  step = 1L,
  behavior = "drag-tap",
  continuous_update = TRUE,
  description = "",
  description_allow_html = FALSE,
  disabled = FALSE,
  orientation = "horizontal",
  readout = TRUE,
  readout_format = "d",
  style = SliderStyle(),
  ...,
  error_call = current_env()
)
}
\arguments{
\item{value}{Int value}

\item{min}{Min value}

\item{max}{Max value}

\item{step}{Minimum step to increment the value}

\item{behavior}{Slider dragging behavior.}

\item{continuous_update}{Update the value of the widget as the user is holding the slider.}

\item{description}{Description of the control.}

\item{description_allow_html}{Accept HTML in the description.}

\item{disabled}{Enable or disable user changes}

\item{orientation}{Vertical or horizontal.}

\item{readout}{Display the current value of the slider next to it.}

\item{readout_format}{Format for the readout}

\item{style}{Must inherit from \link{jupyter.widget.SliderStyle}.}

\item{...}{forwarded to \link[jupyter.widgets.base:DOMWidget]{jupyter.widgets.base::jupyter.widget.DOMWidget} constructor}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
IntSlider widget
}
