% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntSlider.R
\name{IntSlider}
\alias{IntSlider}
\title{IntSlider widget}
\usage{
IntSlider(..., error_call = current_env())
}
\arguments{
\item{...}{See constructor for \code{jupyter.widgets.IntSlider}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a \link{jupyter.widget.IntSlider} object
}
\description{
IntSlider widget
}
