% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntSlider.R
\name{IntSliderStyle}
\alias{IntSliderStyle}
\title{Style for the IntSlider widget}
\usage{
IntSliderStyle(
  description_width = "",
  handle_color = NULL,
  `_model_module` = "@jupyter-widgets/controls",
  `_model_module_version` = "2.0.0",
  `_model_name` = "IntSliderModel",
  `_view_module` = "@jupyter-widgets/base",
  `_view_count` = NULL,
  `_view_module_version` = "2.0.0",
  `_view_name` = "IntSliderView",
  ...,
  error_call = current_env()
)
}
\arguments{
\item{description_width}{description width}

\item{handle_color}{css color for the handle}

\item{_model_module}{The namespace of the model.}

\item{_model_module_version}{A semver requirement for namespace version containing the model.}

\item{_model_name}{model name}

\item{_view_module}{view mmodule}

\item{_view_count}{view count}

\item{_view_module_version}{view module version}

\item{_view_name}{view name}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Style for the IntSlider widget
}
